+++ src/PHPCR/Shell/Console/Command/Phpcr/VersionRestoreCommand.php
@@ -20,50 +20,49 @@ class VersionRestoreCommand extends BasePhpcrCommand
         $this->setHelp(<<<HERE
 Attempt to restore an old version of a node.
 
-<em>If <info>path</info> is given and <info>versionName</info> is a version name:</em>
+* If <info>path</info> is given and <info>versionName</info> is a version name:
  Restores the node at <info>path</info> to the state defined by the version with
  the specified version name (<info>versionName</info>).
  This method will work regardless of whether the node at path is
  checked-in or not.
 
 
-<em>If <info>path</info> is given and <info>versionName</info> is a VersionInterface instance:
-</em>
- Restores the specified version to <info>path</info>. There must be no existing
- node at <info>path</info>. If one exists, a VersionException is thrown.
- There must be a parent node to the location at <info>path</info>, otherwise a
- PathNotFoundException is thrown.
- If the would-be parent of the location <info>path</info> is actually a property,
- or if a node type restriction would be violated, then a
- ConstraintViolationException is thrown.
+* If <info>path</info> is given and <info>versionName</info> is a VersionInterface instance:
+  Restores the specified version to <info>path</info>. There must be no existing
+  node at <info>path</info>. If one exists, a VersionException is thrown.
+  There must be a parent node to the location at <info>path</info>, otherwise a
+  PathNotFoundException is thrown.
+  If the would-be parent of the location <info>path</info> is actually a property,
+  or if a node type restriction would be violated, then a
+  ConstraintViolationException is thrown.
 
 
-<em>If <info>versionName</info> is VersionInterface instance:</em>
- Restores the node in the current workspace that is the versionable node
- of the specified version to the state reflected in that version.
- This method ignores checked-in status.
+* If <info>versionName</info> is VersionInterface instance:
+  Restores the node in the current workspace that is the versionable node
+  of the specified version to the state reflected in that version.
+  This method ignores checked-in status.
 
 
-<em>If <info>versionName</info> is an array of VersionInterface instances:</em>
- Restores a set of versions at once. Used in cases where a "chicken and
- egg" problem of mutually referring REFERENCE properties would prevent
- the restore in any serial order.
- The following restrictions apply to the set of versions specified: If S
- is the set of versions being restored simultaneously,
- - For every version V in S that corresponds to a missing node, there
-   must also be a parent of V in S.
- - S must contain at least one version that corresponds to an existing
-   node in the workspace.
- - No V in S can be a root version (jcr:rootVersion).
- If any of these restrictions does not hold, the restore will fail
- because the system will be unable to determine the path locations to
- which one or more versions are to be restored. In this case a
- VersionException is thrown.
- The versionable nodes in the current workspace that correspond to the
- versions being restored define a set of (one or more) subgraphs.
+* If <info>versionName</info> is an array of VersionInterface instances:
+  Restores a set of versions at once. Used in cases where a "chicken and
+  egg" problem of mutually referring REFERENCE properties would prevent
+  the restore in any serial order.
+  The following restrictions apply to the set of versions specified: If S
+  is the set of versions being restored simultaneously,
+  - For every version V in S that corresponds to a missing node, there
+    must also be a parent of V in S.
+  - S must contain at least one version that corresponds to an existing
+    node in the workspace.
+  - No V in S can be a root version (jcr:rootVersion).
+  If any of these restrictions does not hold, the restore will fail
+  because the system will be unable to determine the path locations to
+  which one or more versions are to be restored. In this case a
+  VersionException is thrown.
+  The versionable nodes in the current workspace that correspond to the
+  versions being restored define a set of (one or more) subgraphs.
 
-<em>If the restore succeeds the changes made are dispatched immediately;
-</em>
+If the restore succeeds the changes made are dispatched immediately;
+
 there is no need to call save.
 
 If an array of VersionInterface instances is restored, an identifier
@@ -85,8 +84,10 @@ particular, conflicts involving subnodes of the restored node that have
 OnParentVersion settings of COPY or VERSION are also governed by the
 <info>removeExisting</info> flag.
 
-<b>Note:</b> The Java API defines this with multiple differing
-signatures, you need to act accordingly in your implementation.
+Note: 
+
+    The Java API defines this with multiple differing
+    signatures, you need to act accordingly in your implementation.
 HERE
         );
     }